Also double intermediate results for FixEE2017 case

+++ ext/include/JMESystematicsCalculators.h
@@ -133,7 +133,7 @@ public:
 protected:
   std::unique_ptr<FactorizedJetCorrectorCalculator,jetcorrdeleter> m_jetCorrectorProd;
   std::unique_ptr<FactorizedJetCorrectorCalculator,jetcorrdeleter> m_jetCorrectorL1Prod;
-  std::array<float,4> calculateFixEE2017Offset(ROOT::VecOps::RVec<bool>& jet_mask,
+  std::array<double,4> calculateFixEE2017Offset(ROOT::VecOps::RVec<bool>& jet_mask,
       const p4compv_t& jet_pt, const p4compv_t& jet_eta, const p4compv_t& jet_phi, const p4compv_t& jet_mass,
       const p4compv_t& jet_rawcorr, const p4compv_t& jet_area, const p4compv_t& jet_muonSubtrFactor,
       const float rho) const;
